code/ai_tutor/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
click==8.1.7
distro==1.9.0
dnspython==1.16.0
fastapi==0.109.2
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
motor==3.2.0
mypy-extensions==1.0.0
openai==1.12.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pydantic==2.6.1
pydantic_core==2.16.2
pymongo==4.6.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.36.3
tqdm==4.66.1
typing_extensions==4.9.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
